Refactor creation logic into init modules
