+++ src/ingestion/config/__init__.py
@@ -1,0 +1,9 @@
+''' Init module for configuration. '''
+
+from cfg_environ.config import Config, ConfigFacade
+
+
+def get_config() -> Config:
+  config = ConfigFacade('ingestion/config', 'development', 'default')
+
+  return config
